# [FIX] stock: improve _merge_quants query performances