Change the `dupes` CTE in _merge_quants, replacing HAVING
by a WHERE clause, allowing to filter out rows beforehand.

The idea is to use the `COUNT()` function as a  window function over the same partition as the group by
below. This will set count = partition size, i.e. count > 1 => duplicates.

##### speed-up

The speed-up is correlated to the size of the `stock.quant` table as well as the
duplicates "density".

In a customer DB with 30M stock.quant and no duplicates, the `dupes` CTE goes from 1m18s to 54s.
Inventory > Reporting > Inventory Report goes from 1m40 -> 58s.

In a test DB with 100k stock_quants, avg time of the whole query while increasing the duplicates density

| Ratio of duplicates | Before PR | After PR |
|:-------------------:|:----------:|:---------:|
| 0.1% | 310 ms | 218ms |
| 1% | 310 ms | 227 ms |
| 10% | 340 ms | 268 ms |
| 25 % | 382 ms | 330 ms |
| 50 % | 435 ms | 485 ms |
| 99%  | 623 ms | 732 ms|

Test DB with ratio of duplicates = 10%, increasing size of `stock.quant` table.

|  stock.quant rows | Before PR | After PR |
|:------------------:|:----------:|:---------:|
|  10k | 63 ms | 88 ms |
|  100k | 340 ms | 268 ms |
|   1M  | 3.3 s  |  2.7 s |

Test DB with ratio of duplicates = 50%, increasing size of `stock.quant` table.

|  stock.quant rows | Before PR | After PR |
|:------------------:|:----------:|:---------:|
|  10k | 56 ms | 80 ms |
|  100k | 435 ms | 485 ms |
|   1M  | 5.2s  |  5.3s |

Overall, the highest speedup is achieved when the `stock.quant` table size is big while the ratio of duplicates is slow. There are instances where the code before PR is faster. But, given that there is a daily cron that merges stock.quant and that
`_quant_tasks` is called each time the quant view/inventory reports are opened, I think it is safe to assume that 
most of the time, the database will fall into the big `stock.quant`/low duplicates ratio case. 

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr