@amoyaux 2021-06-28
Wait before merge @ryv-odoo is double checking