@ryv-odoo 2021-07-12
Hello @Aurelienvd 
With the last change, it is still a improvement ? The new where clause is really faster than `HAVING count(id) > 1` ?