@Aurelienvd 2021-07-14
Hello @ryv-odoo 
Well to be honest there is something a bit weird with the speed-up report. Actually the last change didn't really impact the speed but for an unknown reason, the base code got faster when I benchmarked it again (even though it's an averaged time with cache invalidations in between). I think I'll benchmark it again while being extra careful on the conditions and then we can decide on what the do.

Also, weirdly enough the time taken by the Inventory Report in the customer DB did not change for base code even though it was faster on a testing DB 🤕 

And when I say faster, I do mean that. The average time of the query in a 1M stock.quant DB with 10% dups dropped from 4.5 to 3.3s for base code. Even though I ran the same code with the same postgres config, same indexes, dropping postgres cache in between to be sure. It's the first time I had such a huge speed gap while doing a benchmark again.