@Aurelienvd 2022-03-11
@odoo/rd-logistics Hello, could someone review this fairly old PR ? To give a bit of context, a customer faced
an issue when accessing Inventory Report concurrently. They had 30M stock.quants in their db. The initial plan
was to optimize the `_merge_quants` method of `stock.quant`. After some benchmarks, however, it was decided
that the performance gains were to small to bother (cf. ryv's last comment).

After discussing with ryv on discord a bit, back when he was in logistics, we decided to add an `ir.config_parameter` to
turn off the merging of quants when opening quant views. The reasoning was that it's a bit overkill to merge quants
each time we access the quant view. The procurement run scheduler cron should be enough to handle most cases, and
this config_parameter could be turned off if a specific db has lots of duplicated quants.

Regarding the commit about SKIP LOCKED, it was ryv's idea to only lock duplicated quants. The `test_merge_quant` was
added following a suggestion of amoyaux (cf. above), although I'm not sure if it's still needed now that
the test `test_unpack_and_quants_merging` has been added.

Thank you 😃