@amoyaux 2022-03-22
robodoo r-
I'm testing with a lock and it seems to create big inconsistencies. Let me check a litle bit deeper